% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simReg.R
\name{acceptance_traces}
\alias{acceptance_traces}
\title{Get boolean vectors indicating whether proposed values in \code{sim_reg} Markov chain were accepted}
\usage{
acceptance_traces(samples)
}
\arguments{
\item{samples}{\code{sim_reg_samples} object, as returned by \code{\link{sim_reg}}.}
}
\description{
Get boolean vectors indicating whether proposed values in \code{sim_reg} Markov chain were accepted
}

