% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as.data.frame.sim_reg_summary}
\alias{as.data.frame.sim_reg_summary}
\title{Convert \code{sim_reg_summary} object to \code{data.frame}}
\usage{
\method{as.data.frame}{sim_reg_summary}(x, ...)
}
\arguments{
\item{x}{Object of class `sim_reg_summary`.}

\item{...}{Non-used arguments.}
}
\description{
Convert \code{sim_reg_summary} object to \code{data.frame}
}

