% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simReg.R
\name{final_sample}
\alias{final_sample}
\title{Get final sample from \code{sim_reg_samples} object}
\usage{
final_sample(samples)
}
\arguments{
\item{samples}{\code{sim_reg_samples} object, as returned by \code{\link{sim_reg}}.}
}
\description{
Get final sample from \code{sim_reg_samples} object
}

