% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{full_names}
\alias{full_names}
\title{Replace term ID names of vector/array with full term names}
\usage{
full_names(ontology, x)
}
\arguments{
\item{ontology}{\code{ontology_index} object.}

\item{x}{Vector/array named by term IDs}
}
\description{
Replace term ID names of vector/array with full term names
}

