% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{get_term_marginals}
\alias{get_term_marginals}
\title{Calculate marginal probability of terms inclusion in \code{phi} from \code{sim_reg_out} object}
\usage{
get_term_marginals(sim_reg_out)
}
\arguments{
\item{sim_reg_out}{Object of class \code{sim_reg_output}.}
}
\value{
Numeric vector of probabilities, named by term ID.
}
\description{
Calculate marginal probability of terms inclusion in \code{phi} from \code{sim_reg_out} object
}

