% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{get_terms}
\alias{get_terms}
\title{Get full set of terms to use in inference procedure based on similarity function arguments}
\usage{
get_terms(args)
}
\arguments{
\item{args}{Named list of named arguments which gets passed to ontological similarity function by \code{sim_reg}.}
}
\value{
Character vector of term IDs.
}
\description{
Get full set of terms to use in inference procedure based on similarity function arguments
}

