% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{leaf.term.pair.marginal.matrix}
\alias{leaf.term.pair.marginal.matrix}
\title{Get matrix of presence of term pairs}
\usage{
leaf.term.pair.marginal.matrix(term.descendancy.matrix, phi.trace,
  terms.to.use = NULL, n = 10)
}
\arguments{
\item{term.descendancy.matrix}{Logical term descendancy matrix, dimensions symmetrically labelled by terms, and where by a cell value of TRUE indicates that the row is the ancestor of the column term (in the sense of the DAG structure of the HPO}

\item{phi.trace}{Delta trace output from \code{\link{sim.reg}} function, where the rows are the values of phi at each iteration}

\item{terms.to.use}{Specify the terms to appear in the plot}

\item{n}{Maximum number of terms to include in the matrix}
}
\value{
Matrix of term-pair frequencies (where by cell i,j contains the frequency of inclusion in phi for term pair i,j)
}
\description{
Get matrix of presence of term pairs
}

