% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{log_BF}
\alias{log_BF}
\alias{log_BF.default}
\alias{log_BF.sim_reg_output}
\title{Calculate log Bayes factor for similarity the model, \code{gamma=1} and baseline model, \code{gamma=0}.}
\usage{
log_BF(x, ...)

\method{log_BF}{default}(x, ...)

\method{log_BF}{sim_reg_output}(x, ...)
}
\arguments{
\item{x}{\code{list} of term sets or \code{sim_reg_output} object.}

\item{...}{If x is a \code{list} term sets, other arguments to pass to \code{\link{sim_reg}}, otherwise this is not used.}
}
\value{
Numeric value.
}
\description{
Calculate log Bayes factor for similarity the model, \code{gamma=1} and baseline model, \code{gamma=0}.
}

