% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{P(gamma=1)}
\alias{P(gamma=1)}
\title{Extract the estimated posterior probability of gamma = 1}
\usage{
"P(gamma=1)"(x)
}
\arguments{
\item{x}{A `sim_reg_samples` or `sim_reg_summary` object}
}
\value{
Numeric value
}
\description{
Extract the estimated posterior probability of gamma = 1
}

