% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.sim_reg_samples}
\alias{plot.sim_reg_samples}
\title{Plot SimReg output}
\usage{
\method{plot}{sim_reg_samples}(x, ontology, ...)
}
\arguments{
\item{x}{Output of call to \code{sim_reg}.}

\item{ontology}{\code{ontology_index} object.}

\item{...}{Non-used arguments.}
}
\value{
Plots
}
\description{
Plot the output of the call to the \code{sim_reg} function. Output contains plots of marginal probabilities of terms and pairs of terms being present phi and histograms giving the distributions of variables. Note, a large device is required for a successful call.
}

