% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{plot.sim_reg_summary}
\alias{plot.sim_reg_summary}
\alias{plot.sim_reg_output}
\title{Plot summary of \code{sim_reg_output} object}
\usage{
\method{plot}{sim_reg_summary}(x, ...)

\method{plot}{sim_reg_output}(x, ...)
}
\arguments{
\item{x}{Object of class \code{sim_reg_summary}.}

\item{...}{Additional arguments to pass to \code{\link{plot_term_marginals}}.}
}
\description{
Plot summary of \code{sim_reg_output} object
}
