% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_term_marginals}
\alias{plot_term_marginals}
\title{Create ontological plot of marginal probabilities of terms}
\usage{
plot_term_marginals(ontology, term_marginals, max_terms = 10,
  min_probability = 0.01, ...)
}
\arguments{
\item{ontology}{\code{ontology_index} object.}

\item{term_marginals}{Numeric vector of marginal probabilities of inclusion in \code{phi} for individual terms, named by the term IDs.}

\item{max_terms}{Maximum number of terms to include in plot. Note that additional terms may be included when terms have the same marginal probability, and common ancestor terms are included.}

\item{min_probability}{Threshold probability of inclusion in \code{phi} for triggering inclusion in plot.}

\item{...}{Additional arguments to pass to \code{onto_plot}}
}
\description{
Create ontological plot of marginal probabilities of terms
}

