% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{print.sim_reg_output}
\alias{print.sim_reg_output}
\title{Print \code{sim_reg_output} object}
\usage{
\method{print}{sim_reg_output}(x, ...)
}
\arguments{
\item{x}{Object of class \code{sim_reg_output}.}

\item{...}{Non-used arguments.}
}
\description{
Print \code{sim_reg_output} object
}
