% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.sim_reg_samples}
\alias{print.sim_reg_samples}
\title{Print \code{sim_reg_samples} object}
\usage{
\method{print}{sim_reg_samples}(x, width = getOption("width"),
  ontology = NULL, ...)
}
\arguments{
\item{x}{Object of class \code{sim_reg_samples}.}

\item{width}{Integer value for width of output}

\item{ontology}{Optional \code{ontology_index} class object allowing additional information about terms sampled in \code{phi} to be printed.}

\item{...}{Non-used arguments.}
}
\description{
Print \code{sim_reg_samples} object
}

