% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{print.sim_reg_summary}
\alias{print.sim_reg_summary}
\title{Print \code{sim_reg_summary} object}
\usage{
\method{print}{sim_reg_summary}(x, ...)
}
\arguments{
\item{x}{Object of class \code{sim_reg_summary}.}

\item{...}{Non-used arguments.}
}
\description{
Print \code{sim_reg_summary} object
}
