% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{prob_association}
\alias{prob_association}
\title{Calculate probability of association between \code{y} and \code{x}}
\usage{
prob_association(..., prior = 0.05)
}
\arguments{
\item{...}{Arguments to pass to \code{\link{log_BF}}.}

\item{prior}{Numeric value determing prior probability that \code{gamma=1}.}
}
\value{
Numeric value.
}
\description{
Calculate probability of association between \code{y} and \code{x}
}

