% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{prune_sim_mat}
\alias{prune_sim_mat}
\title{Prune between-term similarity matrix}
\usage{
prune_sim_mat(ontology, term_sim_mat)
}
\arguments{
\item{ontology}{\code{ontology_index} object.}

\item{term_sim_mat}{Numeric matrix with rows and columns corresponding to (and named by) term IDs, and cells containing the similarity between the row and column term}
}
\value{
Numeric matrix of term-term similarities
}
\description{
Prune similarity of more specific terms to 0
}

