% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{s_phi}
\alias{s_phi}
\title{Evaluate function s_phi}
\usage{
s_phi(...)
}
\arguments{
\item{...}{Arguments to pass to \code{\link{s}}}
}
\value{
Similarities
}
\description{
Evaluate function s_phi
}
\seealso{
\code{\link{s}}
}

