% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{s_x}
\alias{s_x}
\title{Evaluate function s_x}
\usage{
s_x(...)
}
\arguments{
\item{...}{Arguments to pass to \code{\link{s}}}
}
\value{
Similarities
}
\description{
Evaluate function s_x
}
\seealso{
\code{\link{s}}
}

