% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{sim.reg.full.summary}
\alias{sim.reg.full.summary}
\title{Create grid of SimReg output plots}
\usage{
sim.reg.full.summary(hpo.terms,
  term.descendancy.matrix = get.term.descendancy.matrix(hpo.terms,
  unique(as.character(sim.reg.out$phi))), file.name, sim.reg.out)
}
\arguments{
\item{hpo.terms}{R-Object representation of HPO}

\item{term.descendancy.matrix}{Logical term descendancy matrix, dimensions symmetrically labelled by terms, and where by a cell value of TRUE indicates that the row is the ancestor of the column term (in the sense of the DAG structure of the HPO}

\item{file.name}{File to write plots to}

\item{sim.reg.out}{Output of call to \code{sim.reg}}
}
\value{
Plots graph to file
}
\description{
Create a PDF containing a comprehensive set of summary plots for the output of the \code{sim.reg} application. Output contains plots of marginal probabilities of terms and pairs of terms being present phi, traces of variables and their log likelihoods recorded during the MCMC application, and histograms giving the distributions of variables.
}

