% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sim_reg_summary}
\alias{sim_reg_summary}
\title{Create grid of SimReg output plots}
\usage{
sim_reg_summary(ontology, file_name, sim_reg_out)
}
\arguments{
\item{ontology}{\code{ontology_index} object.}

\item{file_name}{File to write plots to}

\item{sim_reg_out}{Output of call to \code{sim_reg}}
}
\value{
Plots graph to file
}
\description{
Create a PDF containing a set of summary plots for the output of the \code{sim_reg} application. Output contains plots of marginal probabilities of terms and pairs of terms being present phi and histograms giving the distributions of variables.
}

