\name{SimReg-package}
\alias{SimReg-package}
\alias{SimReg}
\docType{package}
\title{
Functions for HPO logic regression
}
\description{
Functions for logic regression style algorithms, regression genotype against ontologically encoded phenotypes
}
\details{
\tabular{ll}{
Package: \tab hpoPlot\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-10-01\cr
License: \tab What license is it under?\cr
}
Key functions include hpo.sim.reg, for performing HPO regression of binary response variable (typically representing presence of genetic variant) against ontologically encoded phenotype
}
\author{
Daniel Greene <dg333@cam.ac.uk>

Maintainer: Daniel Greene <dg333@cam.ac.uk>
}
%\references{
%~~ Literature or other references for background information ~~
%}
\keyword{ HPO }
