% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{subset_sim_reg_samples}
\alias{subset_sim_reg_samples}
\title{Subset \code{sim_reg_samples} object}
\usage{
subset_sim_reg_samples(samples, at)
}
\arguments{
\item{samples}{\code{sim_reg_samples} classed object.}

\item{at}{Integer vector of indices of iterations to extract from samples.}
}
\description{
Subset \code{sim_reg_samples} object
}

