% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{sum_log_probs}
\alias{sum_log_probs}
\title{Calculate sum of log probabilities on log scale without over/under-flow}
\usage{
sum_log_probs(log_probs)
}
\arguments{
\item{log_probs}{Numeric vector of probabilities on log scale.}
}
\value{
Numeric value on log scale.
}
\description{
Calculate sum of log probabilities on log scale without over/under-flow
}
