% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{summary.sim_reg_output}
\alias{summary.sim_reg_output}
\title{Get summary of \code{sim_reg_output} object}
\usage{
\method{summary}{sim_reg_output}(object, prior = 0.05, ...)
}
\arguments{
\item{object}{Object of class \code{sim_reg_output}.}

\item{prior}{Prior probability of association.}

\item{...}{Non-used arguments.}
}
\description{
Get summary of \code{sim_reg_output} object
}

