% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.sim_reg_samples}
\alias{summary.sim_reg_samples}
\title{Get summary of \code{sim_reg_samples} object.}
\usage{
\method{summary}{sim_reg_samples}(object, ...)
}
\arguments{
\item{object}{Object of class \code{sim_reg_samples}.}

\item{...}{Non-used arguments}
}
\description{
Get summary of \code{sim_reg_samples} object.
}

