% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{term_marginals}
\alias{term_marginals}
\title{Get marginal frequencies of single terms in phi}
\usage{
term_marginals(phi)
}
\arguments{
\item{phi}{List of character vectors of term IDs, as sampled by \code{sim_reg} procedure.}
}
\value{
Numeric vector of proportion of samples each term was included in phi.
}
\description{
Get marginal frequencies of single terms in phi
}

