% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{term_marginals_from_phi_vec}
\alias{term_marginals_from_phi_vec}
\title{Get marginal frequencies of single terms in phi}
\usage{
term_marginals_from_phi_vec(term_descendancy_matrix, phi_vector_trace)
}
\arguments{
\item{term_descendancy_matrix}{Logical term descendancy matrix, dimensions symmetrically labelled by terms, and where by a cell value of TRUE indicates that the row is the ancestor of the column term (in the sense of the DAG structure of the HPO.}

\item{phi_vector_trace}{Delta trace output from \code{\link{sim_reg}} function, where the rows are the values of phi at each iteration}
}
\value{
Numeric vector of proportion of samples each term was included in phi.
}
\description{
Get marginal frequencies of single terms in phi
}

