% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{term_pair_marginals}
\alias{term_pair_marginals}
\title{Get matrix of presence of term pairs}
\usage{
term_pair_marginals(phi, terms_to_use = NULL, symmetric = FALSE, n = 10)
}
\arguments{
\item{phi}{List of character vectors of term IDs, as sampled by \code{sim_reg} procedure.}

\item{terms_to_use}{Specify the terms to appear in the plot.}

\item{symmetric}{Logical value determining whether output is a symmetric matrix (\code{TRUE}) or lower triangular one (\code{FALSE}).}

\item{n}{Maximum number of terms to include in the matrix.}
}
\value{
Matrix of term-pair frequencies (where by cell i,j contains the frequency of inclusion in phi for term pair i,j).
}
\description{
Get matrix of presence of term pairs
}

