% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{term_pair_marginals_plot}
\alias{term_pair_marginals_plot}
\title{Plot marginal distribution of term-pair inclusion in phi}
\usage{
term_pair_marginals_plot(x, ...)
}
\arguments{
\item{x}{Numeric matrix.}

\item{...}{Other parameters to pass to image.}
}
\value{
Nothing - side effect: plots graph.
}
\description{
Plots the marginal frequency of terms pairs' inclusion in the phi parameter over the course of an application of \code{\link{sim_reg}} as heatmap.
}

