% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot.sim_reg_summary}
\alias{plot.sim_reg_summary}
\title{Plot \code{sim_reg_summary} object}
\usage{
\method{plot}{sim_reg_summary}(x, ontology, max_plot_terms = 10, cex = 1,
  ...)
}
\arguments{
\item{x}{Object of class \code{sim_reg_summary}.}

\item{ontology}{\code{ontology_index} object.}

\item{max_plot_terms}{Maximum number of terms (with non-zero sample frequency) to show in plot.}

\item{cex}{Text size for plot.}

\item{...}{Other arguments to pass to \code{onto_plot}.}
}
\description{
Plot \code{sim_reg_summary} object
}

