% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.sim_reg_summary}
\alias{print.sim_reg_summary}
\title{Print \code{sim_reg_summary} object}
\usage{
\method{print}{sim_reg_summary}(x, width = getOption("width"),
  ontology = NULL, ...)
}
\arguments{
\item{x}{Object of class \code{sim_reg_summary}.}

\item{width}{Integer value for width of output}

\item{ontology}{Optional `ontology_index` class object allowing additional information about terms sampled in `phi` to be printed.}

\item{...}{Non-used arguments.}
}
\description{
Print \code{sim_reg_summary} object
}

