% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{term_marginals}
\alias{term_marginals}
\title{Calculate marginal probability of terms inclusion in \code{phi}}
\usage{
term_marginals(...)
}
\arguments{
\item{...}{Arguments to pass to \code{\link{sim_reg}}.}
}
\value{
Numeric vector of probabilities, named by term ID.
}
\description{
Calculate marginal probability of terms inclusion in \code{phi}
}
